fsAdmin/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
pillow
django-cors-headers
openpyxl